refactor helpers into utils
