+++ src/vivarium_testing_utils/automated_validation/data_loader.py
@@ -14,8 +14,8 @@ from vivarium_testing_utils.automated_validation.data_transformation.calculation
 from vivarium_testing_utils.automated_validation.data_transformation.data_schema import (
     SimOutputData,
     SingleNumericColumn,
-    check_io,
 )
+from vivarium_testing_utils.automated_validation.data_transformation.utils import check_io
 
 
 class DataSource(Enum):
