+++ src/vivarium_testing_utils/automated_validation/data_transformation/data_schema.py
@@ -58,35 +58,3 @@ class RatioData(pa.DataFrameModel):
         # and both are numeric
         numeric_columns = df.select_dtypes(include=["number"]).columns
         return len(numeric_columns) == df.shape[1] == 2
-
-
-def check_io(**model_dict):
-    """
-    A wrapper for pa.check_io that automatically converts SchemaModels to schemas.
-
-    Args:
-        **model_dict: Keyword arguments where keys are parameter names and values
-                     are SchemaModel classes or schema objects.
-
-    Returns:
-        A decorator function that wraps the target function with pa.check_io.
-    """
-    # Convert any SchemaModel classes to schemas
-    schema_dict = {}
-    for key, value in model_dict.items():
-        # Check if it's a SchemaModel class (not instance) and has to_schema method
-        if hasattr(value, "to_schema") and callable(value.to_schema):
-            schema_dict[key] = value.to_schema()
-        else:
-            # If it's already a schema or something else, use it as is
-            schema_dict[key] = value
-
-    # Return the decorator using pa.check_io with the converted schemas
-    return pa.check_io(**schema_dict)
-
-
-# TODO: Remove this function and references when we can support Series schemas
-# more easily
-def series_to_dataframe(series: pd.Series[float]) -> pd.DataFrame:
-    """Convert a Series to a DataFrame with the Series values as a single column."""
-    return series.to_frame(name="value")
