+++ src/vivarium_testing_utils/automated_validation/data_transformation/measures.py
@@ -12,13 +12,13 @@ from vivarium_testing_utils.automated_validation.data_transformation.data_schema
     RatioData,
     SimOutputData,
     SingleNumericColumn,
-    check_io,
 )
 from vivarium_testing_utils.automated_validation.data_transformation.formatting import (
     PersonTime,
     SimDataFormatter,
     TransitionCounts,
 )
+from vivarium_testing_utils.automated_validation.data_transformation.utils import check_io
 
 
 class Measure(ABC):
